fix complexity issues
