+++ gedbrowser-api/src/main/java/org/schoellerfamily/gedbrowser/api/transformers/ApiModelToGedObjectVisitor.java
@@ -1,5 +1,7 @@
 package org.schoellerfamily.gedbrowser.api.transformers;
 
+import java.util.List;
+
 import org.schoellerfamily.gedbrowser.api.datamodel.ApiAttribute;
 import org.schoellerfamily.gedbrowser.api.datamodel.ApiFamily;
 import org.schoellerfamily.gedbrowser.api.datamodel.ApiHead;
@@ -123,27 +125,21 @@ public final class ApiModelToGedObjectVisitor implements ApiObjectVisitor {
      * @param apiParent the parent object
      */
     private void addAttributes(final ApiPerson apiParent) {
-        for (final ApiObject object : apiParent.getAttributes()) {
-            final ApiModelToGedObjectVisitor visitor = createVisitor();
-            object.accept(visitor);
-        }
-        for (final ApiObject object : apiParent.getImages()) {
-            final ApiModelToGedObjectVisitor visitor = createVisitor();
-            object.accept(visitor);
-        }
-        for (final ApiObject object : apiParent.getFamc()) {
-            final ApiModelToGedObjectVisitor visitor = createVisitor();
-            object.accept(visitor);
-        }
-        for (final ApiObject object : apiParent.getFams()) {
-            final ApiModelToGedObjectVisitor visitor = createVisitor();
-            object.accept(visitor);
-        }
-        for (final ApiObject object : apiParent.getRefn()) {
-            final ApiModelToGedObjectVisitor visitor = createVisitor();
-            object.accept(visitor);
-        }
-        for (final ApiObject object : apiParent.getChanged()) {
+        addToAttributes(apiParent.getAttributes());
+        addToAttributes(apiParent.getImages());
+        addToAttributes(apiParent.getFamc());
+        addToAttributes(apiParent.getFams());
+        addToAttributes(apiParent.getRefn());
+        addToAttributes(apiParent.getChanged());
+    }
+
+    /**
+     * Add to the attributes of the parent for saving.
+     *
+     * @param attributes some attributes to add to the list
+     */
+    private void addToAttributes(final List<ApiAttribute> attributes) {
+        for (final ApiObject object : attributes) {
             final ApiModelToGedObjectVisitor visitor = createVisitor();
             object.accept(visitor);
         }
